Upgrade to AssemblyScript v0.23.x (#10)
* Add test output to .gitignore

The test output is a build artifact that shouldn't be tracked by Git,
and it changes depending on the AssemblyScript version. Hence, this
commit adds it to .gitignore.

* Update AssemblyScript and handle arrow functions

This commit updates the uses of *Kind to match with AssemblyScript,
which now uses Pascal case. In addition, arrow functions are now
handled properly by as-covers to fix failing tests.